dbtools/requirements.txt
mysql-connector-python==8.3.0
dify_plugin~=0.0.1b66
redis>=5.0.0
